Eliminate 52 more type: ignore suppressions with proper type narrowing
Phase 2 of type safety enforcement. Fixes across 29 files:

- var-annotated: Add explicit type annotations to 7 untyped list variables
- return-value: Use bool()/float()/np.asarray() for numpy scalar returns
  instead of suppressing (biometric, graph_based, concept_drift, gosnn)
- no-any-return: Replace np.clip suppresses with np.asarray() wrappers
  (calibration, conformal_prediction)
- assignment: Fix numpy/Python type inconsistencies in if/else branches
  by converting numpy scalars at source (gosnn_3r, predictive_coding,
  multi_agent_coordination, adaptive_domain_thresholding, score_calibration)
- operator: Convert numpy scalars before builtin operations (iris_recognition,
  fingerprint_recognition, realworld_benchmark, harmonics, acceleration_dynamics)
- Proper None-guard assertions for Optional typed fields (enhanced_statistical)

Total reduction: 611 -> 437 type: ignore comments (174 removed, 28.5%)
Remaining 437 are genuinely necessary: optional dependency imports (39),
torch/framework operations (120+), numpy array creation patterns (80+),
and framework-specific type stub limitations.

mypy: Success (416 source files, 0 errors)
Tests: 5080 passed, 0 regressions

+++ src/omni_mercury_engine/biometric/iris_recognition.py
@@ -712,6 +712,6 @@ class IrisRecognizer:
 
         contrast = np.std(normalized_iris[mask])
 
-        quality = 0.4 * usable_ratio + 0.3 * min(1.0, sharpness * 10) + 0.3 * min(1.0, contrast * 5)  # type: ignore[operator]
+        quality = 0.4 * usable_ratio + 0.3 * min(1.0, float(sharpness * 10)) + 0.3 * min(1.0, float(contrast * 5))
 
         return float(quality)
